+++ validation/graylog_rest_api.py
@@ -191,7 +191,7 @@ class GraylogRestApi:
         }
         return self._create_alert_rule(title, stream, 'AND', time, additional_threshold_type='<', additional_threshold=additional_threshold, second_stream=second_stream)
 
-    def create_alert_rule_or(self, title, time, description):
+    def create_alert_rule_or(self, title, time, description, group_by_fields=[]):
         rule = {
             'field': 'a',
             'type': 1,
@@ -211,7 +211,7 @@ class GraylogRestApi:
             'field_rule': [rule],
             'matching_type': 'AND'
         }
-        return self._create_alert_rule(title, stream, 'OR', time, description=description, additional_threshold_type='<', second_stream=second_stream)
+        return self._create_alert_rule(title, stream, 'OR', time, description=description, additional_threshold_type='<', second_stream=second_stream, group_by_fields=group_by_fields)
 
     def get_alert_rule(self, name):
         response = self._get(f'plugins/com.airbus_cyber_security.graylog.wizard/alerts/{name}')